+++ modules/refinements.js
@@ -7,6 +7,9 @@ module.exports = {
   Imagepath: t.refinement(t.String, (i) => /.(jpg|jpeg|gif|png|svg)$/.test(i), 'Image file'),
   Handle: t.refinement(t.String, (i) => /^@((?!(:|\\|\/)).)*$/.test(i), 'Handle'),
   Date: t.refinement(t.String, (d) => moment(d, moment.ISO_8601, true).isValid(), 'ISO 8601 Date'),
+  MsDate: t.refinement(t.Number, (d) => !isNaN(new Date(d).getTime()), 'Date in ms since January 1, 1970'),
+  DigitString: t.refinement(t.String, (s) => /^\d+$/.test(s), 'String with digits only'),
+  Email: t.refinement(t.String, (s) => s.contains('@'), 'Email'),
 
   /**
    * Ensures that value does not exceed specified length