+++ CHANGELOG.md
@@ -22,6 +22,13 @@
 
 - [static][templates] Added `favicon-16x16.png` and `favicon-32x32.png`. Some browsers like Chrome refusing to use `favicon.ico` in their stead.
 - [package] Added NPM scripts `lint:scripts` and `lint:styles` to quickly invoke linting of scripts or styles-only.
+
+- [modules] Added new `tcomb` refinements:
+
+   * `MsDate` — expect timestamp in ms
+   * `DigitString` — expect strings with digits only
+   * `Email` — expect valid email
+
 - [tests] Added better Jest watch experience with [`jest-watch-typeahead`](https://github.com/jest-community/jest-watch-typeahead).
 
 ### Changed